// gen-types.ts
export type Maybe<T> = T | null;
// ...
export type Agent = Entity & Node & {
  __typename?: 'Agent';
  agentImage: Asset;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description: RichText;
  /** Get the document in other stages */
  documentInStages: Array<Agent>;
  email: Scalars['String']['output'];
  firstName: Scalars['String']['output'];
  /** List of Agent versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  lastName: Scalars['String']['output'];
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Agent>;
  phoneNumber: Scalars['String']['output'];
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  realEstate: Array<RealEstate>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Aggregate = {
  __typename?: 'Aggregate';
  count: Scalars['Int']['output'];
};

/** Asset system model */
export type Asset = Entity & Node & {
  __typename?: 'Asset';
  alt?: Maybe<Scalars['String']['output']>;
  coverImageBlogPost: Array<BlogPost>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<Asset>;
  /** The file name */
  fileName: Scalars['String']['output'];
  /** The file handle */
  handle: Scalars['String']['output'];
  /** The height of the file */
  height?: Maybe<Scalars['Float']['output']>;
  /** List of Asset versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  imageAgent: Array<Agent>;
  imageRealEstate: Array<RealEstate>;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Asset>;
  /** The mime type of the file */
  mimeType?: Maybe<Scalars['String']['output']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  /** The file size */
  size?: Maybe<Scalars['Float']['output']>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  /** Get the url for the asset with provided transformations applied. */
  url: Scalars['String']['output'];
  /** The file width */
  width?: Maybe<Scalars['Float']['output']>;
};
// ...
export type BlogPost = Entity & Node & {
  __typename?: 'BlogPost';
  content: RichText;
  coverImage: Asset;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<BlogPost>;
  excerpt: Scalars['String']['output'];
  /** List of BlogPost versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<BlogPost>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  /** related blog posts */
  relatedBlogPosts: Array<BlogPost>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  title: Scalars['String']['output'];
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type BlogPostConnection = {
  __typename?: 'BlogPostConnection';
  aggregate: Aggregate;
  /** A list of edges. */
  edges: Array<BlogPostEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export enum Districts {
  Bemowo = 'Bemowo',
  Bialoleka = 'Bialoleka',
  Bielany = 'Bielany',
  Mokotow = 'Mokotow',
  Ochota = 'Ochota',
  Remebertow = 'Remebertow',
  Targowek = 'Targowek',
  Ursus = 'Ursus',
  Ursynow = 'Ursynow',
  Wawer = 'Wawer',
  Wesola = 'Wesola',
  Wilanow = 'Wilanow',
  Wlochy = 'Wlochy',
  Wola = 'Wola',
  Zoliborz = 'Zoliborz'
}
// ...
export type Query = {
  __typename?: 'Query';
  /** Retrieve a single agent */
  agent?: Maybe<Agent>;
  /** Retrieve document version */
  agentVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple agents */
  agents: Array<Agent>;
  /** Retrieve multiple agents using the Relay connection interface */
  agentsConnection: AgentConnection;
  /** Retrieve a single asset */
  asset?: Maybe<Asset>;
  /** Retrieve document version */
  assetVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple assets */
  assets: Array<Asset>;
  /** Retrieve multiple assets using the Relay connection interface */
  assetsConnection: AssetConnection;
  /** Retrieve a single blogPost */
  blogPost?: Maybe<BlogPost>;
  /** Retrieve document version */
  blogPostVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple blogPosts */
  blogPosts: Array<BlogPost>;
  /** Retrieve multiple blogPosts using the Relay connection interface */
  blogPostsConnection: BlogPostConnection;
  /** Fetches an object given its ID */
  entities?: Maybe<Array<Entity>>;
  /** Fetches an object given its ID */
  node?: Maybe<Node>;
  /** Retrieve a single realEstate */
  realEstate?: Maybe<RealEstate>;
  /** Retrieve document version */
  realEstateVersion?: Maybe<DocumentVersion>;
  /** Retrieve multiple realEstates */
  realEstates: Array<RealEstate>;
  /** Retrieve multiple realEstates using the Relay connection interface */
  realEstatesConnection: RealEstateConnection;
  /** Retrieve a single scheduledOperation */
  scheduledOperation?: Maybe<ScheduledOperation>;
  /** Retrieve multiple scheduledOperations */
  scheduledOperations: Array<ScheduledOperation>;
  /** Retrieve multiple scheduledOperations using the Relay connection interface */
  scheduledOperationsConnection: ScheduledOperationConnection;
  /** Retrieve a single scheduledRelease */
  scheduledRelease?: Maybe<ScheduledRelease>;
  /** Retrieve multiple scheduledReleases */
  scheduledReleases: Array<ScheduledRelease>;
  /** Retrieve multiple scheduledReleases using the Relay connection interface */
  scheduledReleasesConnection: ScheduledReleaseConnection;
  /** Retrieve a single user */
  user?: Maybe<User>;
  /** Retrieve multiple users */
  users: Array<User>;
  /** Retrieve multiple users using the Relay connection interface */
  usersConnection: UserConnection;
};
// ...
export type RealEstate = Entity & Node & {
  __typename?: 'RealEstate';
  agent?: Maybe<Agent>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description: RichText;
  districtName: Districts;
  /** Get the document in other stages */
  documentInStages: Array<RealEstate>;
  /** List of RealEstate versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  numberOfRooms: Scalars['Int']['output'];
  price: Scalars['Int']['output'];
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  realEstateImage: Asset;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  squareFootage: Scalars['Int']['output'];
  /** System stage field */
  stage: Stage;
  transaction: Transaction;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  yearOfConstruction: Scalars['Int']['output'];
};
// ...
export type RichText = {
  __typename?: 'RichText';
  /** Returns HTMl representation */
  html: Scalars['String']['output'];
  /** Returns Markdown representation */
  markdown: Scalars['String']['output'];
  /** Returns AST representation */
  raw: Scalars['RichTextAST']['output'];
  /** Returns plain-text contents of RichText */
  text: Scalars['String']['output'];
};
// ...
export enum Transaction {
  Rent = 'rent',
  Sale = 'sale'
}
// ...
export type BlogPostsQuery = { __typename?: 'Query', blogPosts: Array<{ __typename?: 'BlogPost', slug: string, title: string, excerpt: string, coverImage: { __typename?: 'Asset', alt?: string | null, url: string } }> };

export type BlogPostQueryVariables = Exact<{
  slug: Scalars['String']['input'];
}>;


export type BlogPostQuery = { __typename?: 'Query', blogPosts: Array<{ __typename?: 'BlogPost', title: string, content: { __typename?: 'RichText', html: string }, coverImage: { __typename?: 'Asset', url: string, alt?: string | null }, relatedBlogPosts: Array<{ __typename?: 'BlogPost', slug: string, excerpt: string, title: string, coverImage: { __typename?: 'Asset', url: string, alt?: string | null } }> }> };

export type BlogPostCountQueryVariables = Exact<{ [key: string]: never; }>;


export type BlogPostCountQuery = { __typename?: 'Query', blogPostsConnection: { __typename?: 'BlogPostConnection', aggregate: { __typename?: 'Aggregate', count: number } } };

export type RealEstatesQueryVariables = Exact<{
  minSquareFootage: Scalars['Int']['input'];
  maxSquareFootage: Scalars['Int']['input'];
  transaction: Array<InputMaybe<Transaction>> | InputMaybe<Transaction>;
  maxPrice: Scalars['Int']['input'];
  minPrice: Scalars['Int']['input'];
  locations: Array<InputMaybe<Districts>> | InputMaybe<Districts>;
  skip: Scalars['Int']['input'];
}>;


export type RealEstatesQuery = { __typename?: 'Query', realEstates: Array<{ __typename?: 'RealEstate', districtName: Districts, price: number, squareFootage: number, slug: string, realEstateImage: { __typename?: 'Asset', url: string, alt?: string | null } }>, realEstatesConnection: { __typename?: 'RealEstateConnection', aggregate: { __typename?: 'Aggregate', count: number } } };

export type AdvertisementQueryVariables = Exact<{
  slug: Scalars['String']['input'];
}>;


export type AdvertisementQuery = { __typename?: 'Query', realEstate?: { __typename?: 'RealEstate', districtName: Districts, numberOfRooms: number, price: number, yearOfConstruction: number, transaction: Transaction, squareFootage: number, description: { __typename?: 'RichText', html: string }, agent?: { __typename?: 'Agent', email: string, firstName: string, phoneNumber: string, lastName: string, slug: string, agentImage: { __typename?: 'Asset', url: string } } | null, realEstateImage: { __typename?: 'Asset', url: string, alt?: string | null } } | null };

export type DistrictsQueryVariables = Exact<{ [key: string]: never; }>;


export type DistrictsQuery = { __typename?: 'Query', __type?: { __typename?: '__Type', enumValues?: Array<{ __typename?: '__EnumValue', name: string }> | null } | null };

export type AgentsQueryVariables = Exact<{ [key: string]: never; }>;


export type AgentsQuery = { __typename?: 'Query', agents: Array<{ __typename?: 'Agent', email: string, firstName: string, lastName: string, slug: string, phoneNumber: string, agentImage: { __typename?: 'Asset', url: string } }> };

export type AgentQueryVariables = Exact<{
  slug: Scalars['String']['input'];
}>;


export type AgentQuery = { __typename?: 'Query', agent?: { __typename?: 'Agent', email: string, firstName: string, lastName: string, phoneNumber: string, description: { __typename?: 'RichText', html: string }, agentImage: { __typename?: 'Asset', url: string }, realEstate: Array<{ __typename?: 'RealEstate', districtName: Districts, price: number, squareFootage: number, slug: string, realEstateImage: { __typename?: 'Asset', alt?: string | null, url: string } }> } | null };

export type HomepageQueryVariables = Exact<{ [key: string]: never; }>;


export type HomepageQuery = { __typename?: 'Query', realEstates: Array<{ __typename?: 'RealEstate', districtName: Districts, price: number, squareFootage: number, slug: string, realEstateImage: { __typename?: 'Asset', alt?: string | null, url: string } }>, blogPosts: Array<{ __typename?: 'BlogPost', slug: string, title: string, excerpt: string, coverImage: { __typename?: 'Asset', url: string, alt?: string | null } }> };